13-1: Manual & unit test code for sfmr g_c_d_id()
Manual test code and Check unit test code for get_container_device_id().
Tests are passing.
Valgrind is happy.

SPOILERS: This is the first of the unit tests that *actually* tests
against a socket file.  Now I have to double back and fix that unit test
for all the legacy unit test files as well.

Changes to be committed:
	new file:   code/test/check_sfmr_get_container_device_id.c
	new file:   code/test/test_sfmr_get_container_device_id.c