13-2: Initial commit of set_ownership() unit tests
Still not done but I've been sitting on this draft for a while.  No
longer.

I refactored the test fixture setup and teardown to use a new struct I
defined, populate that struct, and teardown those structs.

I think that's done.  The source compiles (even if all the test cases
are commented out) so I think the next step is to write run_test_case().
The first step in that process though is to design and write the
what-is-the-expected-result-for-this-input function (that compares the
permissions of the calling process to the expected result of setting the
UID/GID)

EOD COMMIT

Changes to be committed:
	new file:   code/test/check_sfmw_set_ownership.c