13-4: Updated build system for skid_file_link
I don't have any manual test code or unit tests yet... but at least the
Makefile appears to be ready to receive.

TO DO: DON'T DO NOW...
- Write some manual test code for both functions
- Write some unit tests

NOTE: Be sure to test some of the odd edge cases from the man page.
Consider getting in front of the "destination won't be overwritten" but
prematurely returning an errno value (if the default behavior isn't
good and/or silently fails).

Changes to be committed:
	modified:   code/Makefile_linux