dev: Removed clear cmds from Makefile
I discovered the cause of the weird escape characters in the output
file.
PLOT TWIST: The clear command clears the screen by printing characters?!
See: https://stackoverflow.com/a/37778152

The clear commands aren't really necessary anymore anyway.  The output
is so long it scrolls off the screen.

Changes to be committed:
	modified:   Makefile

Changes not staged for commit:
	modified:   README.md
	modified:   code/Makefile

Untracked files:
	devops/