dev: Initial commit of 13-2 export script
I wrote a script to help automate the process of extricating output into
a single file to upload, for my mentor, as output proof.

This is good enough.
Maybe, next time I copy/paste/update this shell script... I implement an
optional CLI argument to redirect all output to.  (Or maybe not)

Changes to be committed:
	new file:   devops/scripts/13-2_export.sh

Untracked files:
	devops/files/