13-4: Initial commit of 13-4 export script
It's not done.
It's mostly commented.
I just wanted to commit it before I forgot to also commit the execute
bits.

My manual test copy/paste commands are working but this seemed like a
great time to translate those into a working copy of the export shell
script for this work.

Changes to be committed:
	new file:   devops/scripts/13-4_export.sh

Changes not staged for commit:
	modified:   code/include/skid_debug.h
	modified:   code/test/test_sfmw_set_ownership.c
	modified:   devops/scripts/13-2_export.sh
	modified:   devops/scripts/13-4_export.sh

Untracked files:
	code/test/test_sfl_create_sym_link.c