13-4: Updated SKID Debug macros
Added errno value, as a decimal, to PRINT_ERRNO().
Then, I noticed the decimal values were getting out of control so I
added some "line" context to the line number value in the output.
It appears to be improved.  It should also be backwards compatible.

Changes to be committed:
	modified:   code/include/skid_debug.h

Changes not staged for commit:
	modified:   code/test/test_sfmw_set_ownership.c
	modified:   devops/scripts/13-2_export.sh
	modified:   devops/scripts/13-4_export.sh

Untracked files:
	code/test/test_sfl_create_sym_link.c