13-6: Initial commit of SKID's file_ops library
Initial commit of the file operations library.
I haven't compiled it yet, much less used it.
I haven't decided if I'm going to write any unit tests for it (because
file I/O is pretty thrashy).
But here it is.

TO DO: DON'T DO NOW (and maybe not every)...
* Consider unit tests for the public functions
* Add a read_file() (after you've implemented the SKID memory library)

Changes to be committed:
	new file:   code/include/skid_file_operations.h
	new file:   code/src/skid_file_operations.c