13-6: Initial commit of SKID's memory library
Hasn't been compiled.
No idea if it works (thought it should).

I will not be unit testing this library.
That's what I hired Valgrind to do.

Meanwhile, time to double-back to the file_operations library and
implement a read file function (now that the SKID libraries have a
standard method of allocating and free'ing heap memory).

Then, get to work on implementing the dir_operations library (see: last
weeks' looooooong TO DO list).

Changes to be committed:
	new file:   code/include/skid_memory.h
	new file:   code/src/skid_memory.c