13-6: Implemented new s_f_m_r lib f_t_t() function
Implemented the new format_time_terse() function

NOTE: This is an uncharacteristic copy/paste (from format_time()) on my
part.  I decided against writing a function that took a format string as
an argument... for safety reasons.  Even if it was local to the library!

Barring that, there wasn't any additional benefit to doing so, so I
didn't.

"Let history be my judge."
Changes to be committed:
	modified:   code/src/skid_file_metadata_read.c

Untracked files:
	code/test/check_sdo_create_dir.c