13-6: Fixed devops_code resolve_to_repo() BUG
Fixed a BUG in a feature I've only just started using.
I wanted resolve_to_repo() to also be able to just give me the absolute
path of the SKID repo, but hadn't tried it yet.

I used that feature, found a BUG, and fixed it.

Legacy unit test cases are all passing.
Valgrind is happy.
Still working on the create_dir() unit tests though.

Changes to be committed:
	modified:   code/src/devops_code.c

Changes not staged for commit:
	modified:   code/test/check_sdo_create_dir.c