SKID-8: gcc now happy w/ simple stream client code
I now realize the need to write a send_socket() (and maybe even a
transmit_socket() wrapper) to mirror recv_socket() (and
receive_socket()).

But that's not what I'm working on at the moment.  So I'll let just this
code ride, as is, until I'm ready to extricate the send() functionality
I already wrote into a SKID library function.

As of now, this client code compiles... and that's enough.

Changes to be committed:
	modified:   code/test/test_sn_simple_stream_client.c

Changes not staged for commit:
	modified:   code/include/skid_network.h
	modified:   code/src/skid_network.c