SKID-8: Cleaned up manual test simple stream code
Transitioned the stream client completely to write_fd() (because we'll
use send_socket() for the DGRAM code).

Added a bit of formatting to some of the "The call was..." error logging
to make it look less like DEBUGGING and more like the rest of the error
feedback.

Code compiles.
Server/client still working.

I guess it's time to see what Valgrind has to say about it all.

Changes to be committed:
	modified:   code/test/test_sn_simple_stream_client.c
	modified:   code/test/test_sn_simple_stream_server.c