SKID-8: Updated manual test simple stream server
Now, the server only accepts one message before the server exits.
I put this in as a temp shunt to help verify Valgrind is happy (to avoid
a SIGINT) but I decided to leave it in.

I'll do something fancier with the next manual test example server:
DGRAM.  Maybe I'll take client input from the CLI, send it, and then
receive a message from the server.  Something like that...

Changes to be committed:
	modified:   code/test/test_sn_simple_stream_server.c