SKID-8: Fixed(?) man test dgram server code
Fixed manual test simple dgram server BUG.
Post-bind(), I was checking the server protocol against the server type.
That's no good.  Fixed.

I also improved DEBUGGING output (after implementing a call to
resolve_protocol()) and refactored the flow of execution a bit.

At this point, at least it's listening for the client...
So that's something...

Changes to be committed:
	modified:   code/test/test_sn_simple_dgram_server.c