SKID-8: Working man test dgram client/server code
GOOD NEWS: This server is now receiving data from this client.
(PLOT TWIST: It's even working locally!)

BAD NEWS: This version of the code highlighted a couple mysteries
that need to be solved.

1. What is it about recv_from_socket() that's blocking but recvfrom()
is working?
2. Why isn't recvfrom() configuring the cliaddr variable with details
about the client the message was received from?
(The DEBUGGING statement in skid_network's get_inet_addr() confirms
the sa_family is coming in as zero (0))

At least this is a positive step in the right direction.

Dirty, interim, EOD commit.

Changes to be committed:
	modified:   code/src/skid_network.c
	modified:   code/test/test_sn_simple_dgram_server.c