SKID-8: Fixed man test dgram server code
I put it in a loop so it can receive multiple connecctions.
Also, it seems that the cause of the "no client address" BUG was I
was giving the pointer to the variable for the size of the sockaddr
struct but it wasn't populated (because I thought it was an "out"
parameter).

The server now accepts, and prints, multiple connections during its
lifetime.

Changes to be committed:
	modified:   code/test/test_sn_simple_dgram_server.c