SKID-8: Fixed Buffer Overflow BUG in s_network.c
The source of the BUG was a combination of check_sn_space() not
accounting for nul-terminator space and strncat truncating
nul-terminators.  Fixed.

Ignoring the DEBUGGING statements (which I've left in for now)...
- Fixed an off-by-one buffer overflow BUG (highlighted by ASAN)
- Updated the call to recvfrom() (it's effectively the same)
- Updated the call to strncat() (not the issue, but I'm leaving it anyway)

Valgrind and ASAN give this code a clean bill of health.

SPOILERS: I still have an issue where I'm reading *everything* that
comes in from the client, so that's next.

Changes to be committed:
	modified:   code/src/skid_network.c

Changes not staged for commit:
	modified:   code/test/test_sn_simple_dgram_client.c
	modified:   code/test/test_sn_simple_dgram_server.c