SKID-8: Updated man test dgram server file comment
Updated the file comment block with the manual compilation command I
used to compile with ASAN (because ASAN isn't easily integrated with the
build system) and the Valgring command I used.

PRO TIP: Use ASAN if the server is doing infi-reads.  Use Valgrind if
it's doing one-and-done reads.

Changes to be committed:
	modified:   code/test/test_sn_simple_dgram_server.c