SKID-8: Updated man test dgram server to "wait"
The entire recv_from_socket() process implicitly does not wait (even if
the do-not-wait flag is missing from flags) so the network module can
determine the size of incoming messages and read them into right-sized
heap-allocated buffers.

As such, the server code had to be refactored to respond to ENODATA
results.

Changes to be committed:
	modified:   code/test/test_sn_simple_dgram_server.c

Changes not staged for commit:
	modified:   code/src/skid_network.c
	modified:   code/test/test_sn_simple_dgram_client.c