SKID-8: Manual dgram testing found a bound
I discovered an upper-end bound in default datagram communication.  I
was gradually increasing the size of the message I was sending to verify
the server would read it all...

...when I received a EMSGSIZE error.  Certainly, I can circumvent
this... but it requries more research.  As such, the dgram client code
is still "instrumented for madness".

Changes to be committed:
	modified:   code/test/test_sn_simple_dgram_client.c

Changes not staged for commit:
	modified:   code/src/skid_network.c