1-3: Added signal code translation to skid_signals
Added the translate_signal_code() function to skid_signals to translate
signal codes into human-readable strings (much like strsignal() does for
signal numbers).

My ability to test it is limited but it appears to be working.
If there's a BUG in it, it's from all the copy/pasting I had to do from
sigaction(2).

Changes to be committed:
	modified:   code/include/skid_signals.h
	modified:   code/src/skid_signals.c

Changes not staged for commit:
	modified:   code/Makefile_linux
	modified:   code/test/test_ssh_handle_interruptions.c

Untracked files:
	code/test/test_ssh_handle_ext_signal_code.c