1-3: Queue int read signal handler manual test IC
Initial commit of manual test code for a parent to read integer values
from a child that is sending sigqueue() messages.

GOOD NEWS: It's working (except not quite as desired).
BAD NEWS: The parent is dropping signals (techinally, overwriting atomic
variable values in a race-conditiony sort of way) without the "tasteful
sleep" in the child process.

TO DO: DON'T DO NOW:
- Finish the fork() implementation
	- verify the child hasn't died (inside the while loop)
	- wait for the child to die after the for loop
	- kill the child(?) if it takes too long
- Write one more extended signal handler
	- Use SIGUSR1 to send integers
	- Use SIGUSR2 to send pointers (pre-fork shared memory space(?))
	- Replicate bi-directional pipes to avoid missed data
- Call it done and submit this work as JQR proof

Changes to be committed:
	new file:   code/test/test_ssh_handle_ext_read_queue_int.c