SKID-12: Intiial commit of Check test framework
I didn't know if it was possible (SPOILERS: It is) but I've been
copy/pasting setup(), teardown(), etc from Check unit test file to file.
No longer.

Time to begin the long and arduous process of fixing this tech debt.

Why now?  Test cases regarding symbolic links are failing.  I think they
were relying on manually created/source-controlled symbolic links.

get_access_time is successfully(?) using the new code on a limited bases
(n07?).  The Makefile recipes have been updated to account for the new
library to link in.

This is the beginning of the war against tech debt...

Changes to be committed:
	new file:   code/include/unit_test_code.h
	new file:   code/src/unit_test_code.c

Changes not staged for commit:
	modified:   code/Makefile_linux
	modified:   code/test/check_sfmr_get_access_time.c
	modified:   code/test/check_sfmr_get_serial_num.c