SKID-12: Build system supports unit test framework
Updated the build system to include the new unit_test_code framework.

- Added variables for unit_test* code
- Removed unecessary DEBUGGING
- Updated the SFMR recipe to include the new unit test framework code
- Updated the default "Linking" echo statement to be more specific

GOOD NEWS: Everything builds (after wrestling with it for a bit)

BAD NEWS: Unit tests are failing (so SKID-12 isn't done yet)

This is part of the remove-tech-debt effort to have a SPOT for the
common filesnames and their creation/cleanup for all the unit tests.

Changes to be committed:
	modified:   code/Makefile_linux

Changes not staged for commit:
	modified:   code/test/check_sfmr_get_access_time.c
	modified:   code/test/check_sfmr_get_serial_num.c