SKID-12: Fixed(?) sym_link.txt issue
First of all, know that the changes here are meaningless.
Only this commit message matters.

FUN FACT: The way git handles Linux symbolic links that are
source-controlled caused problems with some of my unit test cases.  I
suspect that I had manually created the symlink on the original host I
started SKID on.  Then I source-controlled it, but that didn't matter
because it was still a symlink on my host but the git repo converted it
to a flat file.  Yadda yadda yadda... start developing on a fresh
host... yadda yadda yadda... it's not a symlink anymore yadda yadda
yadda test cases failing.

I extricated some common-use libcheck/Check unit test functionality into
a unit_test_code, set the test framework to always delete-and-recreate
the symlink at "test time", and implemented that functionality in the
source file with the failing test cases.  Eureka!  All better.

Changes to be committed:
	typechange: code/test/test_input/sym_link.txt