SKID-12: Time to remove this file from the repo
It's being made as an absolute target... which will now change the
contents for each system.  No one wants to see that.  The setup()
function from unit_test_code will now recreate it for each test case so
it should be fine to remove it.

Changes to be committed:
	deleted:    code/test/test_input/sym_link.txt

Changes not staged for commit:
	modified:   .gitignore
	modified:   code/test/check_sfmr_get_access_time.c
	modified:   code/test/check_sfmr_get_serial_num.c