13-3: Mostly done with redirect_bin_output
Mostly finished.
The child process is now dup2()ing stdout and stderr to the files and
then execvp()ing the bin command, bin args included.
Valgrind is happy.

All that's left to do is to make the parent safely wait for the child to
finish and exit.

Changes to be committed:
	modified:   code/test/redirect_bin_output.c

Untracked files:
	20250211-145927-output_bin-errors.txt
	20250211-145927-output_bin-output.txt
	20250211-145954-output_bin-errors.txt
	20250211-145954-output_bin-output.txt
	20250211-150012-output_bin-errors.txt
	20250211-150012-output_bin-output.txt