13-3: Initial commit of sfc i_c_o_e() test cases
Initial commit of the skid_file_control is_close_on_exec() unit tests.
Only one test case is acutally implemented (the rest are stubbed out)
but normal 1 appears to be passing as is.

TO DO: DON'T DO NOW... implement the remaining test cases and run it all
through Valgrind.

Changes to be committed:
	new file:   code/test/check_sfc_is_close_on_exec.c