13-3: SKID File Descriptors BUG FIX?!
There was a BUG in validate_sfd_args()
The default value of result was EBADF and there was an edge case in
which the default value wasn't being overwritten.
How did this go unnoticed for so long?!

The worst part about finding this BUG was I was working on fcntl()-based
file descriptor locks so I assumed the error was coming from someplace
relevant!

Changes to be committed:
	modified:   code/src/skid_file_descriptors.c

Changes not staged for commit:
	modified:   code/src/skid_file_control.c

Untracked files:
	code/test/test_sfc_read_locked_fd.c