13-3: Initial commit of read_locked_fd() man test
Initial commit of the read_lock_fd() manual test code.
It continuously loops, allowing other current and future manual test
code to operate on that given filename as well.
It appears to be working as expected.
Now it's time to write test_sfc_write_locked_fd.c so they can all fight
with each other.

Changes to be committed:
	new file:   code/test/test_sfc_read_locked_fd.c