11-4: Tweaking query Makefile
I was wrestling with getting this LKM to compile against 5.15.0-135-generic.
It wasn't until I upgraded to 5.15.0-136-generic that I got it to
compile.  Is it the 135 image or was it the act of installing a new
image from "the scratch" that fixed the "Invalid module format" issue I
was having.  The world may never know...

What matters is, I'm back on track and query.ko is working, as-is.

Changes to be committed:
	modified:   modules/query/Makefile

Changes not staged for commit:
	modified:   modules/Makefile
	deleted:    modules/Makefile_constants
	modified:   modules/query/query.c

Untracked files:
	modules/Makefile_module_constants