8-2: Added more preprocessor checks to skid_assemb
Added another error-raising preprocessor check to verify a GNU compiler,
or one compatible with it, is being used (since I'm using GCC
attributes).

I may move these later.  That, or change the logical ifndef behavior to
be a bit more forgiving on non-supported architectures.  (No reason to
throw the baby out with the bath water)

Changes to be committed:
	modified:   code/src/skid_assembly.c