2-5: Initial commit of cleanup attribute man test
Initial commit of the manual test code to highlight the effectiveness of
the gcc attribute "cleanup".

GOOD NEWS: The code works.
BADA NEWS: Valgrind is reporting a memory leak (which shouldn't be the
case if the cleanup attribute was working as advertised).

Now, it's time to troubleshoot...

Changes to be committed:
	new file:   code/test/test_sm_raii_void_macro.c