4-5: Intial commit of manuall mmap() test code
I know this won't compile (because it's not being linked against the
thread library, at the very least) but I wanted an interim commit.

Now, it's time to get the build system working and then brink the manual
test file all the way home.

TO DO: DON'T DO NOW... the parent code block likely needs to wait_pid()
on the child process.

Changes to be committed:
	new file:   code/test/test_sm_mapped_memory.c