4-5: Refactored skid_memory manual test code
Refactored the manual test code to map the memory before
the fork() to ensure the virtual memory address was common
to both the parent and the child.

That meant I had to split some functionality (size/map and
read/memcpy) into different functions.

Also, I changed the data type of data in process_queue_read()
to make utilizing it easier.

GOOD NEWS: Manual test code works as-is.
BAD NEWS: I haven't asked Valgrind about it yet.

Changes to be committed:
	modified:   code/test/test_sm_handle_ext_read_queue_ptr.c