4-2: Final(?) build system file.ext update
I learned, some time ago, that a .mk file extension will tell Sublime
Text 3 "this is a Makefile"... even if it's named something non-standard
(like Makefile_linux).

Easier to update the build system if Sublime can do syntax highlighting.

GOOD NEWS: The build system appears to doing what it was before this
commit.

BAD NEWS: What was it doing?  Failing to link because I haven't added a
rule to link the skid_clone manual test code.

Changes to be committed:
	modified:   Makefile
	modified:   code/Makefile
	renamed:    code/Makefile_constants -> code/Makefile_constants.mk
	renamed:    code/Makefile_linux -> code/Makefile_linux.mk