4-2: Updated clone3 wrapper to set exit signal
My main() usage of the new library wrapper wasn't identifying that any
child process had exited, much less that it even *had* a child process.

Silly me... the child process needed a way to communicate that it
finished and (the easiest way) to do that is to use SIGCHLD.

Updated the implementation and the API comment block in the header.

Changes to be committed:
	modified:   code/include/skid_clone.h
	modified:   code/src/skid_clone.c

Changes not staged for commit:
	modified:   code/test/test_sc_sandbox_process.c