4-2: Copy/pasted script output
The PIDs changed (of course), but the thing to note in this commit diff
is that I removed the duplicate MAIN() output.

Redirecting the shell script into the output.txt file caused that
duplication.  (I'm reasonably assured I narrowed it down through manual
testing)

To clean it up, I copy pasted the output into the text file.

Changes to be committed:
	modified:   devops/files/4-2_output.txt