8-5: Initial commit of nostdlib code
I had to change the name from test_misc_gcc_options because the build
system automatically picks up on source files that start with test_ and
tries to link them.

This source file needs a bespoke Makefile rule to get it where it needs
to go.

The manual command in the file comment block is correct, and works, but
it's missing a few things.

- The comment is stale (it's not testing skid_assembly anymore)
- It needs the code to acutally execute the file
- It needs a bit more of an explanation

This branch also needs
- An export script

Changes to be committed:
	new file:   code/test/showcase_gcc_options.c