4-3: Missing socket file mystery SOLVED
PLOT TWIST: It was the client, in /tmp, with the call to unlink()

TL;DR - The client was calling unlink() on the socket file.
Fixed.

Things are working as they should now.
The server is taking, and logging, all-comers.
The client is just connecting and closing.

Changes to be committed:
	modified:   code/test/test_sn_socket_logging_client.c

Changes not staged for commit:
	modified:   code/src/skid_file_descriptors.c
	modified:   code/src/skid_file_operations.c
	modified:   code/src/skid_network.c
	modified:   code/test/test_sn_socket_logging_server.c