4-3: Added some safety to socket server
Socket server now checks for the socket file in the while() loop.
There's still a chance for an edge case in which the socket file is
unlink()'d after the "blocking" call to accept_client().

Possible solutions:
1. Add a wrapper to accept4() in skid_network (so we can use the
   non-blocking flag)
2. Directly call accept4() in the manual test code (so we can use the
   non-blocking flag)
3. Fork: child process listens, parent process becomes the "environment
   watchdog"
4. Let it ride

SPOILERS: In the interest of time, I'm choosing 4.

Changes to be committed:
	modified:   code/test/test_sn_socket_logging_server.c

Changes not staged for commit:
	modified:   code/src/skid_network.c
	modified:   code/test/test_sn_socket_logging_client.c