4-3: Initial commit of sem func wrappers
Initial commit of a new semaphore library.
Frankly, I wanted to make it a bit easier to separate named sem vs.
unnamed sem vs. any-old sem functionality separate.
This is it.

I haven't compiled this yet, much less tested it.

This is Phase 1 of a big shared mem manual test code refactor.

Changes to be committed:
	new file:   code/include/skid_semaphores.h
	new file:   code/src/skid_semaphores.c

Changes not staged for commit:
	modified:   code/include/skid_macros.h
	modified:   code/include/skid_memory.h
	modified:   code/include/skid_validation.h
	modified:   code/src/skid_validation.c
	modified:   code/test/test_sm_shared_mem.h
	modified:   code/test/test_sm_shared_mem_client.c
	modified:   code/test/test_sm_shared_mem_server.c