7-1: Continued implementing child swarm test code
GOOD NEWS: It's coming along, and this commit compiles...

BAD NEWS: ...but it won't link because the build system doesn't have a
default rule for test_misc_*

SPOILERS: I'm going to rename this to test_sp_ so I can use the
skid_pipe default build rule.  I think it should be fine.  I don't think
I need to do any special linking and I don't expect I'll be compelled to
wrap the poll() (or select()) calls in a library...

(However, that's how I felt about semaphores and we see how *that* went
in 4-3...)

Changes to be committed:
	modified:   code/test/test_misc_child_swarm_poll.c