7-1: Big step for poll()ing manual test code
- Removed DEBUGGING
- Removed stale code
- Fixed read_pollfd() response*

* - Unlike most skid functions, NULL didn't always mean error.
  read_pollfd() returning NULL could be "no data" or, if errnum was set,
  something went wrong.  The no data use case does not count as an
  error.

GOOD NEWS: This code is working as expected.
BAD NEWS: Valgrind hates it.  Why?  The parent is orphaning the
children.  Presumably, everything shutsdown nice and neat, but some of
the children are sleep()ing when the SIGINT signal comes through.  They
awake to find their parent dead (along with most of their siblings?!).
Dark.

Time to add the "wait for the children to die" code block in the
parent's cleanup.

Changes to be committed:
	modified:   code/test/test_sp_child_swarm_poll.c

Changes not staged for commit:
	modified:   code/src/skid_poll.c