7-1: Another step for poll() test code
I kill the children, wait for them to die, and (currently) report the
status.
The good news is, this is working.
The bad news is, the children all have an as-is copy of some of the
allocated memory from the parent.

atexit() won't work... but maybe my SKID_AUTO_FREE_VOID can save the
day.  (I'd hate to refactor to static arrays)

Changes to be committed:
	modified:   code/test/test_sp_child_swarm_poll.c

Changes not staged for commit:
	modified:   code/src/skid_poll.c