11-2: Truncating export script ls output
...for privacy sake.
I *should* be exporting this data from my development VM but I'm not
sure how easy it will be to get my PNY thumb drive through Windows into
a Linux VM... *and* make this script work.

We're rolling with it.  Time to wrap this up.

Changes to be committed:
	modified:   devops/scripts/11-2_export.sh

Untracked files:
	devops/files/11-2_output.txt